# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mango",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/fontes-de-renda": {
      "get": {
        "tags": [
          "fonte_de_renda"
        ],
        "summary": "Listar",
        "operationId": "listar_api_fontes_de_renda_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FonteDeRendaRead"
                  },
                  "type": "array",
                  "title": "Response Listar Api Fontes De Renda Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "fonte_de_renda"
        ],
        "summary": "Criar",
        "operationId": "criar_api_fontes_de_renda_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FonteDeRendaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FonteDeRendaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fontes-de-renda/{fonte_id}": {
      "get": {
        "tags": [
          "fonte_de_renda"
        ],
        "summary": "Obter",
        "operationId": "obter_api_fontes_de_renda__fonte_id__get",
        "parameters": [
          {
            "name": "fonte_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fonte Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FonteDeRendaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "fonte_de_renda"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_fontes_de_renda__fonte_id__patch",
        "parameters": [
          {
            "name": "fonte_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fonte Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FonteDeRendaUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FonteDeRendaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "fonte_de_renda"
        ],
        "summary": "Remover",
        "operationId": "remover_api_fontes_de_renda__fonte_id__delete",
        "parameters": [
          {
            "name": "fonte_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fonte Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/credenciais-pluggy": {
      "get": {
        "tags": [
          "credencial_pluggy"
        ],
        "summary": "Listar",
        "operationId": "listar_api_credenciais_pluggy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CredencialPluggyRead"
                  },
                  "type": "array",
                  "title": "Response Listar Api Credenciais Pluggy Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "credencial_pluggy"
        ],
        "summary": "Criar",
        "operationId": "criar_api_credenciais_pluggy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredencialPluggyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredencialPluggyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/credenciais-pluggy/{credencial_id}": {
      "get": {
        "tags": [
          "credencial_pluggy"
        ],
        "summary": "Obter",
        "operationId": "obter_api_credenciais_pluggy__credencial_id__get",
        "parameters": [
          {
            "name": "credencial_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Credencial Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredencialPluggyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "credencial_pluggy"
        ],
        "summary": "Atualizar",
        "operationId": "atualizar_api_credenciais_pluggy__credencial_id__patch",
        "parameters": [
          {
            "name": "credencial_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Credencial Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredencialPluggyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredencialPluggyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "credencial_pluggy"
        ],
        "summary": "Remover",
        "operationId": "remover_api_credenciais_pluggy__credencial_id__delete",
        "parameters": [
          {
            "name": "credencial_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Credencial Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Obter",
        "operationId": "obter_api_dashboard_get",
        "parameters": [
          {
            "name": "inicio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Inicio"
            }
          },
          {
            "name": "fim",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fim"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResumo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/series": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Series",
        "operationId": "series_api_dashboard_series_get",
        "parameters": [
          {
            "name": "inicio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Inicio"
            }
          },
          {
            "name": "fim",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fim"
            }
          },
          {
            "name": "granularidade",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "diaria",
                "semanal",
                "mensal"
              ],
              "type": "string",
              "default": "semanal",
              "title": "Granularidade"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSeries"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CredencialPluggyCreate": {
        "properties": {
          "client_id": {
            "type": "string",
            "minLength": 1,
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "minLength": 1,
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "client_secret"
        ],
        "title": "CredencialPluggyCreate"
      },
      "CredencialPluggyRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "usuario_id": {
            "type": "integer",
            "title": "Usuario Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          }
        },
        "type": "object",
        "required": [
          "id",
          "usuario_id",
          "criado_em",
          "atualizado_em"
        ],
        "title": "CredencialPluggyRead"
      },
      "CredencialPluggyUpdate": {
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "title": "CredencialPluggyUpdate"
      },
...
      "DashboardResumo": {
        "properties": {
          "saldo_total_centavos": {
            "type": "integer",
            "title": "Saldo Total Centavos"
          },
          "entradas_centavos": {
            "type": "integer",
            "title": "Entradas Centavos"
          },
          "saidas_centavos": {
            "type": "integer",
            "title": "Saidas Centavos"
          },
          "resultado_centavos": {
            "type": "integer",
            "title": "Resultado Centavos"
          },
          "nao_revisadas": {
            "type": "integer",
            "title": "Nao Revisadas"
          },
          "gasto_por_categoria": {
            "items": {
              "$ref": "#/components/schemas/GastoCategoria"
            },
            "type": "array",
            "title": "Gasto Por Categoria"
          },
          "ultimas_transacoes": {
            "items": {
              "$ref": "#/components/schemas/TransacaoRead"
            },
            "type": "array",
            "title": "Ultimas Transacoes"
          },
          "faturas_abertas": {
            "items": {
              "$ref": "#/components/schemas/FaturaRead"
            },
            "type": "array",
            "title": "Faturas Abertas"
          }
        },
        "type": "object",
        "required": [
          "saldo_total_centavos",
          "entradas_centavos",
          "saidas_centavos",
          "resultado_centavos",
          "nao_revisadas",
          "gasto_por_categoria",
          "ultimas_transacoes",
          "faturas_abertas"
        ],
        "title": "DashboardResumo"
      },
      "DashboardSeries": {
        "properties": {
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/SerieBucket"
            },
            "type": "array",
            "title": "Buckets"
          }
        },
        "type": "object",
        "required": [
          "buckets"
        ],
        "title": "DashboardSeries"
      },
...
      "FaturaRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "cartao_id": {
            "type": "integer",
            "title": "Cartao Id"
          },
          "pluggy_bill_id": {
            "type": "string",
            "title": "Pluggy Bill Id"
          },
          "due_date": {
            "type": "string",
            "format": "date-time",
            "title": "Due Date"
          },
          "total_amount_centavos": {
            "type": "integer",
            "title": "Total Amount Centavos"
          },
          "total_amount_currency_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount Currency Code"
          },
          "minimum_payment_centavos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Payment Centavos"
          },
          "allows_installments": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allows Installments"
          },
          "usuario_id": {
            "type": "integer",
            "title": "Usuario Id"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cartao_id",
          "pluggy_bill_id",
          "due_date",
          "total_amount_centavos",
          "usuario_id",
          "criado_em",
          "atualizado_em"
        ],
        "title": "FaturaRead"
      },
...
      "FonteDeRendaCreate": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Nome"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "fixa",
              "variavel"
            ],
            "title": "Tipo"
          },
          "valor_estimado_centavos": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Valor Estimado Centavos"
          },
          "recorrencia": {
            "type": "string",
            "enum": [
              "mensal",
              "trimestral",
              "semestral",
              "anual",
              "irregular"
            ],
            "title": "Recorrencia"
          },
          "fonte": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Fonte"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "tipo",
          "valor_estimado_centavos",
          "recorrencia"
        ],
        "title": "FonteDeRendaCreate"
      },
      "FonteDeRendaRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "usuario_id": {
            "type": "integer",
            "title": "Usuario Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "fixa",
              "variavel"
            ],
            "title": "Tipo"
          },
          "valor_estimado_centavos": {
            "type": "integer",
            "title": "Valor Estimado Centavos"
          },
          "recorrencia": {
            "type": "string",
            "enum": [
              "mensal",
              "trimestral",
              "semestral",
              "anual",
              "irregular"
            ],
            "title": "Recorrencia"
          },
          "fonte": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fonte"
          }
        },
        "type": "object",
        "required": [
          "id",
          "usuario_id",
          "nome",
          "tipo",
          "valor_estimado_centavos",
          "recorrencia",
          "fonte"
        ],
        "title": "FonteDeRendaRead"
      },
      "FonteDeRendaUpdate": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "tipo": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "fixa",
                  "variavel"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tipo"
          },
          "valor_estimado_centavos": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Estimado Centavos"
          },
          "recorrencia": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "mensal",
                  "trimestral",
                  "semestral",
                  "anual",
                  "irregular"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorrencia"
          },
          "fonte": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Fonte"
          }
        },
        "type": "object",
        "title": "FonteDeRendaUpdate",
        "description": "PATCH parcial — só os campos enviados são alterados."
      },
...
      "GastoCategoria": {
        "properties": {
          "categoria_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "total_centavos": {
            "type": "integer",
            "title": "Total Centavos"
          }
        },
        "type": "object",
        "required": [
          "categoria_id",
          "total_centavos"
        ],
        "title": "GastoCategoria"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SerieBucket": {
        "properties": {
          "inicio": {
            "type": "string",
            "format": "date",
            "title": "Inicio"
          },
          "entradas_centavos": {
            "type": "integer",
            "title": "Entradas Centavos"
          },
          "saidas_centavos": {
            "type": "integer",
            "title": "Saidas Centavos"
          },
          "resultado_centavos": {
            "type": "integer",
            "title": "Resultado Centavos"
          },
          "por_categoria": {
            "items": {
              "$ref": "#/components/schemas/GastoCategoria"
            },
            "type": "array",
            "title": "Por Categoria"
          }
        },
        "type": "object",
        "required": [
          "inicio",
          "entradas_centavos",
          "saidas_centavos",
          "resultado_centavos",
          "por_categoria"
        ],
        "title": "SerieBucket",
        "description": "Um período (semana ou mês) da série temporal. `inicio` é a chave; o rótulo é\nformatado no frontend a partir dela e da granularidade."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}